tools/c7n_kube/pyproject.toml
[tool.poetry]
name = "c7n_kube"
version = "0.2.14"
description = "Cloud Custodian - Kubernetes Provider"
readme = "readme.md"
homepage = "https://cloudcustodian.io"
repository = "https://github.com/cloud-custodian/cloud-custodian"
documentation = "https://cloudcustodian.io/docs/"
authors = ["Cloud Custodian Project"]
license = "Apache-2.0"
classifiers = [
   "License :: OSI Approved :: Apache Software License",
   "Topic :: System :: Systems Administration",
   "Topic :: System :: Distributed Computing"
]

[tool.poetry.dependencies]
python = "^3.7"
kubernetes = "^10.0.1"

[tool.poetry.dev-dependencies]
c7n = {path = "../..", develop = true}
pytest = "^6.0.0"
vcrpy = "^4.0.2"

[build-system]
requires = ["poetry>=0.12", "setuptools"]
build-backend = "poetry.masonry.api"
